Fixed cast warning in an example
git-svn-id: http://svn.code.sf.net/p/adevs/code/trunk@147 153d293f-9360-418f-87de-c6d3afdd7922

+++ examples/modelica/main_robot.cpp
@@ -27,7 +27,7 @@ class RobotWithSensor:
 		void init(double* q)
 		{
 			Robot::init(q);
-			k = get_$Ptheta()/mAngle;
+			k = int(get_$Ptheta()/mAngle);
 		}
 		void extra_state_event_funcs(double* z)
 		{
